+++ cloud/server/models/Price.ts
@@ -1,4 +1,4 @@
-import {AbstractModel} from "../code/AbstractModel";
+import {AbstractModel} from "./Contract/AbstractModel";
 import {PriceInterface} from "./PriceInterface";
 
 export class Price extends AbstractModel implements PriceInterface {
