+++ cloud/server/models/Product.ts
@@ -1,4 +1,4 @@
-import {AbstractModel} from "../code/AbstractModel";
+import {AbstractModel} from "./Contract/AbstractModel";
 import {ProductInterface} from "./ProductInterface";
 
 export class Product extends AbstractModel implements ProductInterface {
