+++ cloud/server/models/User.ts
@@ -1,10 +1,9 @@
-import {AbstractModel} from "../code/AbstractModel";
-import SimpleSchema from 'simpl-schema';
+import {AbstractModel} from "./Contract/AbstractModel";
 import {UserInterface} from "./UserInterface";
 import {Role} from "./Role";
 
 export class User extends AbstractModel implements UserInterface {
-  $collection = "users";
+  $collection                       = "users";
   static LICENSE_PERMISSION_SALES   = 'sales';
   static LICENSE_PERMISSION_AGENCY  = 'agency';
   static LICENSE_PERMISSION_USER    = 'user';
